Modified the gitignore file so that it could ignore the wiki.d and the uploads directory on the wiki.
I decided to use pmwiki as my wiki and this is now included in
this commit.  I made changes to it to have it be integrated
with the application and as of this commit, there are some issues
but this is pretty much complete.  I decided to use an existing
wiki instead of creating my own as I really did want to have a full
featured wiki because I am now using the application on a daily basis
and want to make it my primary source of documentation and other stuff
that I want to keep track of.

Note: I am ignoring the 'wiki.d' and 'uploads' directory under the
wiki directory as this is where I'll be storing my content for the
wiki and don't want these files to be under source control.